@@ -14,8 +14,9 @@
         "@babel/preset-env": "^7.7.7",
         "@fortawesome/fontawesome-free": "^5.10.2",
         "@vue/babel-helper-vue-jsx-merge-props": "^1.0.0",
+        "@vue/babel-plugin-jsx": "^1.0.2",
         "@vue/babel-plugin-transform-vue-jsx": "^1.1.2",
-        "@vue/babel-preset-jsx": "^1.1.2",
+        "@vue/compiler-sfc": "^3.0.5",
         "axios": "^0.18.0",
         "babel-loader": "^8.0.6",
         "backpack.css": "^2.0.0",
@@ -28,21 +29,19 @@
         "jquery": "^3.4.1",
         "mini-css-extract-plugin": "^0.8.0",
         "nano-js": "https://github.com/vankizmann/nano-js.git#master",
-        "nano-ui": "https://github.com/vankizmann/nano-ui.git#master",
+        "nano-ui": "https://github.com/vankizmann/nano-ui.git#origin/next",
         "node-sass": "^4.12.0",
         "postcss": "^8.1.1",
         "postcss-inline-svg": "^4.1.0",
         "postcss-loader": "^4.0.3",
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
-        "simplebar": "^5.1.0",
         "style-loader": "^1.1.2",
         "terser-webpack-plugin": "^2.3.5",
         "url-loader": "^2.1.0",
-        "vue": "^2.6.10",
-        "vue-loader": "^15.8.3",
-        "vue-router": "^3.0.2",
-        "vue-template-compiler": "^2.6.10",
+        "vue": "^3.0.5",
+        "vue-loader": "^16.1.2",
+        "vue-router": "^4.0.3",
         "webpack": "^4.44.2",
         "webpack-cli": "^3.3.12"
     }
