@@ -13,10 +13,10 @@
         "@babel/plugin-transform-object-assign": "^7.2.0",
         "@babel/preset-env": "^7.7.7",
         "@vue/babel-helper-vue-jsx-merge-props": "^1.0.0",
+        "@vue/babel-plugin-jsx": "^1.0.2",
         "@vue/babel-plugin-transform-vue-jsx": "^1.1.2",
-        "@vue/babel-preset-jsx": "^1.1.2",
+        "@vue/compiler-sfc": "^3.0.5",
         "babel-loader": "^8.0.6",
-        "backpack.css": "^2.0.0",
         "copy-webpack-plugin": "^5.0.4",
         "core-js": "^3.0.1",
         "cross-env": "^5.1",
@@ -30,13 +30,11 @@
         "postcss-loader": "^4.0.3",
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
-        "simplebar": "^5.1.0",
         "style-loader": "^1.1.2",
         "terser-webpack-plugin": "^2.3.5",
         "url-loader": "^2.1.0",
-        "vue": "^2.6.10",
-        "vue-loader": "^15.8.3",
-        "vue-template-compiler": "^2.6.10",
+        "vue": "^3.0.5",
+        "vue-loader": "^16.1.2",
         "webpack": "^4.44.2",
         "webpack-cli": "^3.3.12"
     }
