@@ -1,48 +1,4 @@
-import Vue from 'vue';
-window.Vue = Vue;
-
-import Nano from "nano-js";
-Nano.install(Vue.prototype);
-
-import VueNano from "nano-ui";
-Vue.use(VueNano, {
-    checked:        'fa fa-check',
-    intermediate:   'fa fa-minus',
-    clock:          'fa fa-clock',
-    calendar:       'fa fa-calendar',
-    times:          'fa fa-times',
-    success:        'fa fa-check-circle',
-    warning:        'fa fa-exclamation-circle',
-    danger:         'fa fa-times-circle',
-    info:           'fa fa-info-circle',
-    angleUp:        'fa fa-angle-up',
-    angleRight:     'fa fa-angle-right',
-    angleDown:      'fa fa-angle-down',
-    angleLeft:      'fa fa-angle-left',
-    arrowUp:        'fa fa-arrow-up',
-    arrowRight:     'fa fa-arrow-right',
-    arrowDown:      'fa fa-arrow-down',
-    arrowLeft:      'fa fa-arrow-left',
-    create:         'fa fa-plus',
-    save:           'fa fa-check',
-    delete:         'fa fa-trash',
-    archive:        'fa fa-archive',
-    action:         'fa fa-cog',
-    preview:        'fa fa-link',
-    activate:       'fa fa-check',
-    deactivate:     'fa fa-times'
-}, {
-    iconPosition: 'before',
-    notifyPosition: 'bottom-end'
-});
-
-import VueRouter from "vue-router";
-Vue.use(VueRouter);
-
-import Axios from "axios";
-Vue.$http = Vue.prototype.$http = Axios;
-
-require('./config/axios');
+import { NanoInstall } from "nano-ui";
 
 import KyoIndex from "./prototypes/KyoIndex";
 window.KyoIndex = KyoIndex;
@@ -50,45 +6,88 @@ window.KyoIndex = KyoIndex;
 import KyoForm from "./prototypes/KyoForm";
 window.KyoForm = KyoForm;
 
-import KyoBackend from "./layout/KyoBackend";
-
-import KyoLayoutMainmenu from "./layout/KyoLayoutMainmenu";
-Vue.component(KyoLayoutMainmenu.name, KyoLayoutMainmenu);
-
-import KyoLayoutSubmenu from "./layout/KyoLayoutSubmenu";
-Vue.component(KyoLayoutSubmenu.name, KyoLayoutSubmenu);
-
-import KyoLayoutLanguage from "./layout/KyoLayoutLanguage";
-Vue.component(KyoLayoutLanguage.name, KyoLayoutLanguage);
-
-import KyoLayoutUserpanel from "./layout/KyoLayoutUserpanel";
-Vue.component(KyoLayoutUserpanel.name, KyoLayoutUserpanel);
-
-import KyoLayoutWebsite from "./layout/KyoLayoutWebsite";
-Vue.component(KyoLayoutWebsite.name, KyoLayoutWebsite);
-
-import KyoWebsiteNode from "./website/KyoWebsiteNode";
-Vue.component(KyoWebsiteNode.name, KyoWebsiteNode);
-
-import KyoTitlebar from "./helpers/KyoTitlebar";
-Vue.component(KyoTitlebar.name, KyoTitlebar);
-
-import KyoTitlebarSearch from "./helpers/KyoTitlebarSearch";
-Vue.component(KyoTitlebarSearch.name, KyoTitlebarSearch);
-
-import KyoFilterbar from "./helpers/KyoFilterbar";
-Vue.component(KyoFilterbar.name, KyoFilterbar);
-
-import KyoDatatable from "./helpers/KyoDatatable";
-Vue.component(KyoDatatable.name, KyoDatatable);
-
-import KyoDashboard from "./pages/dashboard/KyoDashboard";
-Vue.component(KyoDashboard.name, KyoDashboard);
+window.baseTitle = document.title;
 
 
-Nano.Dom.ready(() => {
+// router = router.afterEach((to) => {
+//     Nano.Dom.title(Nano.Locale.trans(Nano.Obj.get(to, 'meta.menu.title')));
+// });
 
-    window.baseTitle = document.title;
+class Kyoto {
+
+    static components = {};
+    static functions = [];
+
+    static component(name, data = null)
+    {
+        if ( data === null ) {
+            return Nano.Obj.get(this.components, name)
+        }
+
+        Nano.Obj.set(this.components, name, data);
+    }
+
+    static register(App)
+    {
+        Nano.Arr.each(this.functions, (callback) => {
+            callback(App);
+        });
+
+        Nano.Obj.each(this.components, (data, name) => {
+            App.component(name, data);
+        });
+    }
+
+    static ready(callback)
+    {
+        Nano.Arr.push(this.functions, callback);
+    }
+
+}
+
+window.Kyoto = Kyoto;
+
+require('./config/axios');
+
+Nano.Dom.ready(function () {
+
+    let App = Vue.createApp(require("./layout/KyoBackend").default);
+    // App = createApp({});
+
+    let KyoLayoutMainmenu = require("./layout/KyoLayoutMainmenu").default;
+    let KyoLayoutSubmenu = require("./layout/KyoLayoutSubmenu").default;
+    let KyoLayoutLanguage = require("./layout/KyoLayoutLanguage").default;
+    let KyoLayoutUserpanel = require("./layout/KyoLayoutUserpanel").default;
+    let KyoLayoutWebsite = require("./layout/KyoLayoutWebsite").default;
+    let KyoWebsiteNode = require("./website/KyoWebsiteNode").default;
+    let KyoTitlebar = require("./helpers/KyoTitlebar").default;
+    let KyoTitlebarSearch = require("./helpers/KyoTitlebarSearch").default;
+    let KyoFilterbar = require("./helpers/KyoFilterbar").default;
+    let KyoDatatable = require("./helpers/KyoDatatable").default;
+
+    let KyoDashboard = require("./pages/dashboard/KyoDashboard").default;
+    App.component(KyoDashboard.name, KyoDashboard);
+
+    App.component(KyoLayoutMainmenu.name, KyoLayoutMainmenu);
+    App.component(KyoLayoutSubmenu.name, KyoLayoutSubmenu);
+    App.component(KyoLayoutLanguage.name, KyoLayoutLanguage);
+    App.component(KyoLayoutUserpanel.name, KyoLayoutUserpanel);
+    App.component(KyoLayoutWebsite.name, KyoLayoutWebsite);
+    App.component(KyoWebsiteNode.name, KyoWebsiteNode);
+    App.component(KyoTitlebar.name, KyoTitlebar);
+    App.component(KyoTitlebarSearch.name, KyoTitlebarSearch);
+    App.component(KyoFilterbar.name, KyoFilterbar);
+    App.component(KyoDatatable.name, KyoDatatable);
+
+    App.config.devtools = true;
+
+    App.use(function (App) {
+        App.config.globalProperties.$http = axios;
+    });
+
+    App.use(function (App) {
+        Kyoto.register(App);
+    });
 
     let routes =[];
 
@@ -101,7 +100,7 @@ Nano.Dom.ready(() => {
         };
 
         if ( menu.option.component ) {
-            data.component = Vue.component(data.name = menu.option.component);
+            data.component = App.component(data.name = menu.option.component);
         }
 
         if ( menu.option.redirect ) {
@@ -111,21 +110,18 @@ Nano.Dom.ready(() => {
         routes.push(data);
     });
 
-    KyoBackend.router = new VueRouter({
-        base: Nano.Obj.get(window, 'backendPaths.base', '/'), mode: 'history', routes
+    let router = VueRouter.createRouter({
+        history: VueRouter.createWebHistory(Nano.Obj.get(window, 'backendPaths.base', '/')),
+        mode: 'history',
+        routes: routes,
     });
 
-    KyoBackend.router.afterEach((to) => {
-        Nano.Dom.title(Nano.Locale.trans(Nano.Obj.get(to, 'meta.menu.title')));
+    App.use(router);
+
+    App.use(function (App) {
+        NanoInstall(App);
     });
 
-    let rootComponent = {};
-
-    if ( Nano.Dom.find('#app').attr('data-root') === 'true' ) {
-        rootComponent = KyoBackend;
-    }
-
-    window.App = new Vue(rootComponent);
-
-    window.App.$mount('#app');
+    App.mount('#app');
 });
+
