@@ -1,11 +1,11 @@
 <template>
     <div class="kyo-datatable">
 
-        <NTable :items="KyoIndex.result.data" :viewport-height="true" :disable-move="true" :remove-node="false" :insert-node="false" :selected.sync="KyoIndex.selected" :expanded.sync="KyoIndex.expanded" :filter-props.sync="KyoIndex.query.filter" :sort-prop.sync="KyoIndex.query.prop" :sort-dir.sync="KyoIndex.query.dir" :item-height="40" :loading-init="0" :allow-drag="KyoIndex.allowDrag" :allow-drop="KyoIndex.allowDrop" :safe-zone="KyoIndex.safeZone" :render-expand="renderExpand" :group="group" :allow-groups="allowGroups" @move="KyoIndex.onMove" @row-dblclick="rowDblclick" :threshold="500">
+        <NTable :items="KyoIndex.result.data" :disable-move="true" v-model:selected="KyoIndex.selected" v-model:expanded="KyoIndex.expanded" v-model:filter="KyoIndex.query.filter" v-model:sort-prop="KyoIndex.query.prop" v-model:sort-dir="KyoIndex.query.dir" :item-height="40" :loading-init="0" :allow-drag="KyoIndex.allowDrag" :allow-drop="KyoIndex.allowDrop" :safezone="KyoIndex.safeZone" :render-expand="renderExpand" :group="group" :allow-groups="allowGroups" @move="KyoIndex.onMove" @row-dblclick="rowDblclick">
             <slot></slot>
         </NTable>
 
-        <NPaginator :page.sync="KyoIndex.query.page" :limit.sync="KyoIndex.query.limit" :total="KyoIndex.result.total" :layout="renderPagination ? ['limit', 'count', 'spacer', 'goto', 'pages'] : ['count']"></NPaginator>
+        <!-- <NPaginator :page.sync="KyoIndex.query.page" :limit.sync="KyoIndex.query.limit" :total="KyoIndex.result.total" :layout="renderPagination ? ['limit', 'count', 'spacer', 'goto', 'pages'] : ['count']"></NPaginator> -->
 
     </div>
 </template>
@@ -51,7 +51,7 @@
             renderPagination: {
                 default()
                 {
-                    return ! this.renderExpand;
+                    return true;
                 },
                 type: [Boolean]
             },
@@ -62,6 +62,7 @@
 
             rowDblclick(...args)
             {
+                console.log('dblclick')
                 this.$emit('row-dblclick', ...args);
             }
 
