@@ -2,7 +2,7 @@ const path = require('path');
 const webpack = require('webpack');
 const TerserPlugin = require('terser-webpack-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const { VueLoaderPlugin } = require('vue-loader');
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 let jsExport = {
@@ -34,7 +34,7 @@ let jsExport = {
     },
     resolve: {
         alias: {
-            'vue$': path.resolve('node_modules/vue/dist/vue.esm.js'),
+            // 'vue$': path.resolve('node_modules/vue/dist/vue.esm-bundler.js'),
             'nano-js$': path.resolve('node_modules/nano-js/dist/nano-js.js'),
             'nano-ui$': path.resolve('node_modules/nano-ui/dist/nano-ui.js'),
         },
