@@ -20,8 +20,8 @@ return [
     ],
 
     'backend' => [
-        'kyoto/language::js/index.js',
-        'kyoto/language::css/index.css'
+        // 'kyoto/language::js/index.js',
+        // 'kyoto/language::css/index.css'
     ]
 
 ];
