@@ -86,123 +86,27 @@
 /************************************************************************/
 /******/ ({
 
-/***/ "./node_modules/@vue/babel-helper-vue-jsx-merge-props/dist/helper.js":
-/*!***************************************************************************!*\
-  !*** ./node_modules/@vue/babel-helper-vue-jsx-merge-props/dist/helper.js ***!
-  \***************************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports, __webpack_require__) {
-
-"use strict";
-eval("function _extends(){return _extends=Object.assign||function(a){for(var b,c=1;c<arguments.length;c++)for(var d in b=arguments[c],b)Object.prototype.hasOwnProperty.call(b,d)&&(a[d]=b[d]);return a},_extends.apply(this,arguments)}var normalMerge=[\"attrs\",\"props\",\"domProps\"],toArrayMerge=[\"class\",\"style\",\"directives\"],functionalMerge=[\"on\",\"nativeOn\"],mergeJsxProps=function(a){return a.reduce(function(c,a){for(var b in a)if(!c[b])c[b]=a[b];else if(-1!==normalMerge.indexOf(b))c[b]=_extends({},c[b],a[b]);else if(-1!==toArrayMerge.indexOf(b)){var d=c[b]instanceof Array?c[b]:[c[b]],e=a[b]instanceof Array?a[b]:[a[b]];c[b]=d.concat(e)}else if(-1!==functionalMerge.indexOf(b)){for(var f in a[b])if(c[b][f]){var g=c[b][f]instanceof Array?c[b][f]:[c[b][f]],h=a[b][f]instanceof Array?a[b][f]:[a[b][f]];c[b][f]=g.concat(h)}else c[b][f]=a[b][f];}else if(\"hook\"==b)for(var i in a[b])c[b][i]=c[b][i]?mergeFn(c[b][i],a[b][i]):a[b][i];else c[b]=a[b];return c},{})},mergeFn=function(a,b){return function(){a&&a.apply(this,arguments),b&&b.apply(this,arguments)}};module.exports=mergeJsxProps;\n\n\n//# sourceURL=webpack:///./node_modules/@vue/babel-helper-vue-jsx-merge-props/dist/helper.js?");
-
-/***/ }),
-
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js&":
-/*!*****************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js& ***!
-  \*****************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js":
+/*!***************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js ***!
+  \***************************************************************************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageCreate',\n  localized: false,\n  urls: {\n    show: '/{locale}/kyoto/page/http/controllers/page/show',\n    store: '/{locale}/kyoto/page/http/controllers/page/store'\n  },\n  beforeMount: function beforeMount() {\n    this.$root.hideWebsite();\n  },\n  beforeDestroy: function beforeDestroy() {\n    this.$root.showWebsite();\n  },\n  methods: {\n    gotoEdit: function gotoEdit(row) {\n      this.$router.push({\n        name: 'KyoPageEdit',\n        params: row.data\n      });\n      console.log(row.data);\n    },\n    gotoIndex: function gotoIndex() {\n      this.$router.push({\n        name: 'KyoPages'\n      });\n    }\n  },\n  \"extends\": window.KyoForm\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPages',\n  localized: true,\n  urls: {\n    index: '/{locale}/kyoto/page/http/controllers/page/index',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  defaults: function defaults() {\n    var query = {\n      page: 1,\n      limit: 25,\n      prop: 'updated_at',\n      dir: 'asc',\n      filter: [],\n      search: '',\n      columns: ['title', 'content']\n    };\n    return {\n      query: query\n    };\n  },\n  \"extends\": window.KyoIndex\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
 
 /***/ }),
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js&":
-/*!***************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js& ***!
-  \***************************************************************************************************************************************************************/
-/*! exports provided: default */
+/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54":
+/*!**************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54 ***!
+  \**************************************************************************************************************************************************************************/
+/*! exports provided: render */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageEdit',\n  localized: true,\n  urls: {\n    show: '/{locale}/kyoto/page/http/controllers/page/show',\n    update: '/{locale}/kyoto/page/http/controllers/page/update',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  beforeMount: function beforeMount() {\n    this.$root.hideWebsite();\n  },\n  beforeDestroy: function beforeDestroy() {\n    this.$root.showWebsite();\n  },\n  methods: {\n    gotoIndex: function gotoIndex() {\n      this.$router.push({\n        name: 'KyoPages'\n      });\n    }\n  },\n  \"extends\": window.KyoForm\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js&":
-/*!************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js& ***!
-  \************************************************************************************************************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n//\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPages',\n  localized: true,\n  urls: {\n    index: '/{locale}/kyoto/page/http/controllers/page/index',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  defaults: function defaults() {\n    var query = {\n      page: 1,\n      limit: 25,\n      prop: 'updated_at',\n      dir: 'asc',\n      filter: [],\n      search: '',\n      columns: ['title', 'content']\n    };\n    return {\n      query: query\n    };\n  },\n  \"extends\": window.KyoIndex\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js&":
-/*!***********************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js& ***!
-  \***********************************************************************************************************************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n//\n//\n//\n//\n//\n//\n//\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoWidgetHeadline',\n  inject: {\n    NDraggableItem: {\n      \"default\": undefined\n    }\n  },\n  computed: {\n    value: function value() {\n      return this.NDraggableItem.veItem.value;\n    }\n  },\n  mounted: function mounted() {\n    this.$watch('value', Nano.Any.debounce(this.NDraggableItem[\"export\"], 100), {\n      deep: true\n    });\n  }\n});\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878&":
-/*!***********************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878& ***!
-  \***********************************************************************************************************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function () {var _vm=this;var _h=_vm.$createElement;var _c=_vm._self._c||_h;return _c('NLoader',{staticClass:\"full-height-child\",attrs:{\"visible\":_vm.load}},[_c('div',{staticClass:\"grid grid--col\"},[_c('KyoTitlebar',{staticClass:\"col--flex-0-0\",attrs:{\"link\":{ name: 'KyoPages' }}},[_c('template',{slot:\"action\"},[_c('NButtonGroup',[_c('NButton',{attrs:{\"type\":\"primary\"},on:{\"click\":_vm.storeItem}},[_vm._v(\"\\n                        \"+_vm._s(_vm.trans('Apply'))+\"\\n                    \")]),_vm._v(\" \"),_c('NButton',{attrs:{\"type\":\"primary\"},on:{\"click\":_vm.storeItemClose}},[_vm._v(\"\\n                        \"+_vm._s(_vm.trans('Save'))+\"\\n                    \")])],1)],1)],2),_vm._v(\" \"),_c('KyoPageWidgets',{staticClass:\"col--flex-1-0\",model:{value:(_vm.result.builder),callback:function ($$v) {_vm.$set(_vm.result, \"builder\", $$v)},expression:\"result.builder\"}},[_c('NForm',{attrs:{\"form\":_vm.result,\"errors\":_vm.errors}},[_c('NFormGroup',{attrs:{\"icon\":\"fa fa-cog\",\"legend\":_vm.trans('Settings')}},[_c('NFormItem',{attrs:{\"label\":_vm.trans('State'),\"prop\":\"state\"}},[_c('NSwitch',{attrs:{\"on-value\":1,\"off-value\":0},model:{value:(_vm.result.state),callback:function ($$v) {_vm.$set(_vm.result, \"state\", $$v)},expression:\"result.state\"}},[_vm._v(\"\\n                            \"+_vm._s(_vm.trans('Item is enabled'))+\"\\n                        \")])],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Hide'),\"prop\":\"hide\"}},[_c('NSwitch',{attrs:{\"on-value\":1,\"off-value\":0},model:{value:(_vm.result.hide),callback:function ($$v) {_vm.$set(_vm.result, \"hide\", $$v)},expression:\"result.hide\"}},[_vm._v(\"\\n                            \"+_vm._s(_vm.trans('Item is hidden'))+\"\\n                        \")])],1)],1),_vm._v(\" \"),_c('NFormGroup',{attrs:{\"icon\":\"fa fa-compass\",\"legend\":_vm.trans('Menu')}},[_c('NFormItem',{attrs:{\"label\":_vm.trans('Title'),\"prop\":\"title\"}},[_c('NInput',{model:{value:(_vm.result.title),callback:function ($$v) {_vm.$set(_vm.result, \"title\", $$v)},expression:\"result.title\"}})],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Slug'),\"prop\":\"slug\"}},[_c('NInput',{model:{value:(_vm.result.slug),callback:function ($$v) {_vm.$set(_vm.result, \"slug\", $$v)},expression:\"result.slug\"}})],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Content'),\"prop\":\"content\"}},[_c('NTextarea',{model:{value:(_vm.result.content),callback:function ($$v) {_vm.$set(_vm.result, \"content\", $$v)},expression:\"result.content\"}})],1)],1)],1)],1)],1)])}\nvar staticRenderFns = []\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a&":
-/*!*********************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a& ***!
-  \*********************************************************************************************************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function () {var _vm=this;var _h=_vm.$createElement;var _c=_vm._self._c||_h;return _c('NLoader',{staticClass:\"full-height-child\",attrs:{\"visible\":_vm.load}},[_c('div',{staticClass:\"grid grid--col\"},[_c('KyoTitlebar',{staticClass:\"col--flex-0-0\",attrs:{\"link\":{ name: 'KyoPages' }},on:{\"delete\":_vm.deleteItem}},[_c('template',{slot:\"action\"},[_c('NButtonGroup',[_c('NButton',{attrs:{\"type\":\"primary\"},on:{\"click\":_vm.updateItem}},[_vm._v(\"\\n                        \"+_vm._s(_vm.trans('Apply'))+\"\\n                    \")]),_vm._v(\" \"),_c('NButton',{attrs:{\"type\":\"primary\"},on:{\"click\":_vm.updateCloseItem}},[_vm._v(\"\\n                        \"+_vm._s(_vm.trans('Save'))+\"\\n                    \")])],1)],1)],2),_vm._v(\" \"),_c('KyoPageWidgets',{staticClass:\"col--flex-1-0\",model:{value:(_vm.result.builder),callback:function ($$v) {_vm.$set(_vm.result, \"builder\", $$v)},expression:\"result.builder\"}},[_c('NForm',{attrs:{\"form\":_vm.result,\"errors\":_vm.errors}},[_c('NFormGroup',{attrs:{\"icon\":\"fa fa-compass\",\"legend\":_vm.trans('Menu')}},[_c('NFormItem',{attrs:{\"label\":_vm.trans('Title'),\"prop\":\"title\"}},[_c('NInput',{model:{value:(_vm.result.title),callback:function ($$v) {_vm.$set(_vm.result, \"title\", $$v)},expression:\"result.title\"}})],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Slug'),\"prop\":\"slug\"}},[_c('NInput',{model:{value:(_vm.result.slug),callback:function ($$v) {_vm.$set(_vm.result, \"slug\", $$v)},expression:\"result.slug\"}})],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Content'),\"prop\":\"content\"}},[_c('NTextarea',{model:{value:(_vm.result.content),callback:function ($$v) {_vm.$set(_vm.result, \"content\", $$v)},expression:\"result.content\"}})],1)],1),_vm._v(\" \"),_c('NFormGroup',{attrs:{\"icon\":\"fa fa-cog\",\"legend\":_vm.trans('Settings')}},[_c('NFormItem',{attrs:{\"label\":_vm.trans('State'),\"prop\":\"state\"}},[_c('NSwitch',{attrs:{\"on-value\":1,\"off-value\":0},model:{value:(_vm.result.state),callback:function ($$v) {_vm.$set(_vm.result, \"state\", $$v)},expression:\"result.state\"}},[_vm._v(\"\\n                            \"+_vm._s(_vm.trans('Item is enabled'))+\"\\n                        \")])],1),_vm._v(\" \"),_c('NFormItem',{attrs:{\"label\":_vm.trans('Hide'),\"prop\":\"hide\"}},[_c('NSwitch',{attrs:{\"on-value\":1,\"off-value\":0},model:{value:(_vm.result.hide),callback:function ($$v) {_vm.$set(_vm.result, \"hide\", $$v)},expression:\"result.hide\"}},[_vm._v(\"\\n                            \"+_vm._s(_vm.trans('Item is hidden'))+\"\\n                        \")])],1)],1)],1)],1)],1)])}\nvar staticRenderFns = []\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c&":
-/*!******************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c& ***!
-  \******************************************************************************************************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function () {var _vm=this;var _h=_vm.$createElement;var _c=_vm._self._c||_h;return _c('NLoader',{staticClass:\"full-height-child\",attrs:{\"visible\":_vm.load}},[_c('div',{staticClass:\"grid grid--col\"},[_c('KyoTitlebar',{staticClass:\"col--flex-0-0\",on:{\"delete\":_vm.deleteItems}},[_c('template',{slot:\"search\"},[_c('KyoTitlebarSearch',{model:{value:(_vm.query.search),callback:function ($$v) {_vm.$set(_vm.query, \"search\", $$v)},expression:\"query.search\"}})],1),_vm._v(\" \"),_c('template',{slot:\"action\"},[_c('NButton',{attrs:{\"type\":\"primary\"},on:{\"click\":function($event){return _vm.$router.push({ name: 'KyoPageCreate' })}}},[_vm._v(\"\\n                    \"+_vm._s(_vm.trans('Create page'))+\"\\n                \")])],1)],2),_vm._v(\" \"),_c('KyoDatatable',{staticClass:\"col--flex-1-1\",attrs:{\"group\":['page', 'transaction']},on:{\"row-dblclick\":_vm.gotoEdit}},[_c('NTableColumn',{attrs:{\"type\":\"boolean\",\"prop\":\"state\",\"label\":_vm.trans('State'),\"sort\":true,\"filter\":true}}),_vm._v(\" \"),_c('NTableColumn',{attrs:{\"type\":\"boolean\",\"prop\":\"hide\",\"label\":_vm.trans('Hide'),\"sort\":true,\"filter\":true}}),_vm._v(\" \"),_c('NTableColumn',{attrs:{\"type\":\"string\",\"prop\":\"title\",\"label\":_vm.trans('Title'),\"fluid\":true,\"sort\":true,\"filter\":true},scopedSlots:_vm._u([{key:\"default\",fn:function(ref){\nvar value = ref.value;\nreturn _c('router-link',{attrs:{\"to\":{ name: 'KyoPageEdit', params: value }}},[_vm._v(_vm._s(value.title))])}}])}),_vm._v(\" \"),_c('NTableColumn',{attrs:{\"type\":\"datetime\",\"prop\":\"updated_at\",\"label\":_vm.trans('Modified'),\"sort\":true,\"filter\":true}}),_vm._v(\" \"),_c('NTableColumn',{attrs:{\"type\":\"datetime\",\"prop\":\"created_at\",\"label\":_vm.trans('Created'),\"sort\":true,\"filter\":true}})],1)],1)])}\nvar staticRenderFns = []\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0&":
-/*!*****************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0& ***!
-  \*****************************************************************************************************************************************************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function () {var _vm=this;var _h=_vm.$createElement;var _c=_vm._self._c||_h;return _c('div',[_c('NFormItem',{attrs:{\"label\":_vm.trans('Title')}},[_c('NInput',{model:{value:(_vm.value.title),callback:function ($$v) {_vm.$set(_vm.value, \"title\", $$v)},expression:\"value.title\"}})],1)],1)}\nvar staticRenderFns = []\n\n\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
-
-/***/ }),
-
-/***/ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js":
-/*!********************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/runtime/componentNormalizer.js ***!
-  \********************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return normalizeComponent; });\n/* globals __VUE_SSR_CONTEXT__ */\n\n// IMPORTANT: Do NOT use ES2015 features in this file (except for modules).\n// This module is a runtime utility for cleaner component module output and will\n// be included in the final webpack user bundle.\n\nfunction normalizeComponent (\n  scriptExports,\n  render,\n  staticRenderFns,\n  functionalTemplate,\n  injectStyles,\n  scopeId,\n  moduleIdentifier, /* server only */\n  shadowMode /* vue-cli only */\n) {\n  // Vue.extend constructor export interop\n  var options = typeof scriptExports === 'function'\n    ? scriptExports.options\n    : scriptExports\n\n  // render functions\n  if (render) {\n    options.render = render\n    options.staticRenderFns = staticRenderFns\n    options._compiled = true\n  }\n\n  // functional template\n  if (functionalTemplate) {\n    options.functional = true\n  }\n\n  // scopedId\n  if (scopeId) {\n    options._scopeId = 'data-v-' + scopeId\n  }\n\n  var hook\n  if (moduleIdentifier) { // server build\n    hook = function (context) {\n      // 2.3 injection\n      context =\n        context || // cached call\n        (this.$vnode && this.$vnode.ssrContext) || // stateful\n        (this.parent && this.parent.$vnode && this.parent.$vnode.ssrContext) // functional\n      // 2.2 with runInNewContext: true\n      if (!context && typeof __VUE_SSR_CONTEXT__ !== 'undefined') {\n        context = __VUE_SSR_CONTEXT__\n      }\n      // inject component styles\n      if (injectStyles) {\n        injectStyles.call(this, context)\n      }\n      // register component module identifier for async chunk inferrence\n      if (context && context._registeredComponents) {\n        context._registeredComponents.add(moduleIdentifier)\n      }\n    }\n    // used by ssr in case component is cached and beforeCreate\n    // never gets called\n    options._ssrRegister = hook\n  } else if (injectStyles) {\n    hook = shadowMode\n      ? function () {\n        injectStyles.call(\n          this,\n          (options.functional ? this.parent : this).$root.$options.shadowRoot\n        )\n      }\n      : injectStyles\n  }\n\n  if (hook) {\n    if (options.functional) {\n      // for template-only hot-reload because in that case the render fn doesn't\n      // go through the normalizer\n      options._injectStyles = hook\n      // register for functional component in vue file\n      var originalRender = options.render\n      options.render = function renderWithStyleInjection (h, context) {\n        hook.call(context)\n        return originalRender(h, context)\n      }\n    } else {\n      // inject component registration as beforeCreate hook\n      var existing = options.beforeCreate\n      options.beforeCreate = existing\n        ? [].concat(existing, hook)\n        : [hook]\n    }\n  }\n\n  return {\n    exports: scriptExports,\n    options: options\n  }\n}\n\n\n//# sourceURL=webpack:///./node_modules/vue-loader/lib/runtime/componentNormalizer.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nconst _hoisted_1 = { class: \"grid grid--col\" }\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_KyoTitlebarSearch = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebarSearch\")\n  const _component_NButton = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\")\n  const _component_KyoTitlebar = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebar\")\n  const _component_NTableColumn = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NTableColumn\")\n  const _component_router_link = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"router-link\")\n  const _component_KyoDatatable = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoDatatable\")\n  const _component_NLoader = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NLoader\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(_component_NLoader, {\n    visible: _ctx.load,\n    class: \"full-height-child\"\n  }, {\n    default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n      Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", _hoisted_1, [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebar, {\n          class: \"col--flex-0-0\",\n          onDelete: _ctx.deleteItems\n        }, {\n          search: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebarSearch, {\n              modelValue: _ctx.query.search,\n              \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => (_ctx.query.search = $event))\n            }, null, 8 /* PROPS */, [\"modelValue\"])\n          ]),\n          action: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n              type: \"primary\",\n              onClick: _cache[2] || (_cache[2] = $event => (_ctx.$router.push({ name: 'KyoPageCreate' })))\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Create page')), 1 /* TEXT */)\n              ]),\n              _: 1 /* STABLE */\n            })\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onDelete\"]),\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoDatatable, {\n          class: \"col--flex-1-1\",\n          onRowDblclick: _ctx.gotoEdit,\n          group: ['page', 'transaction']\n        }, {\n          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"state\",\n              label: _ctx.trans('State'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"hide\",\n              label: _ctx.trans('Hide'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"string\",\n              prop: \"title\",\n              label: _ctx.trans('Title'),\n              fluid: true,\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_router_link, {\n                  to: { name: 'KyoPageEdit', params: _ctx.item }\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])((item) => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(item.title), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"to\"])\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"updated_at\",\n              label: _ctx.trans('Modified'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"created_at\",\n              label: _ctx.trans('Created'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"])\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onRowDblclick\"])\n      ])\n    ]),\n    _: 1 /* STABLE */\n  }, 8 /* PROPS */, [\"visible\"]))\n}\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
 
 /***/ }),
 
@@ -214,174 +118,43 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _pages_KyoPages__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./pages/KyoPages */ \"./resources/js/pages/KyoPages.vue\");\n/* harmony import */ var _pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./pages/KyoPageCreate */ \"./resources/js/pages/KyoPageCreate.vue\");\n/* harmony import */ var _pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./pages/KyoPageEdit */ \"./resources/js/pages/KyoPageEdit.vue\");\n/* harmony import */ var _builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./builder/KyoPageBuilder */ \"./resources/js/builder/KyoPageBuilder.jsx\");\n/* harmony import */ var _widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./widgets/KyoWidgetHeadline */ \"./resources/js/widgets/KyoWidgetHeadline.vue\");\n\n\nvue__WEBPACK_IMPORTED_MODULE_0___default.a.component(_pages_KyoPages__WEBPACK_IMPORTED_MODULE_1__[\"default\"].name, _pages_KyoPages__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\nvue__WEBPACK_IMPORTED_MODULE_0___default.a.component(_pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_2__[\"default\"].name, _pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_2__[\"default\"]);\n\nvue__WEBPACK_IMPORTED_MODULE_0___default.a.component(_pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_3__[\"default\"].name, _pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_3__[\"default\"]);\n\nvue__WEBPACK_IMPORTED_MODULE_0___default.a.component(_builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_4__[\"default\"].name, _builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_4__[\"default\"]);\n\nvue__WEBPACK_IMPORTED_MODULE_0___default.a.component(_widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_5__[\"default\"].name, _widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_5__[\"default\"]);\n\n__webpack_require__(/*! ./configs/page */ \"./resources/js/configs/page.js\");\n\nif (console && console.log) {\n  console.log('kyoto/page ready.');\n}\n\n//# sourceURL=webpack:///./resources/js/bootstrap.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./pages/KyoPages2 */ \"./resources/js/pages/KyoPages2.vue\");\n// import KyoPageCreate from \"./pages/KyoPageCreate\";\n// App.component(KyoPageCreate.name, KyoPageCreate);\n// import KyoPageEdit from \"./pages/KyoPageEdit\";\n// App.component(KyoPageEdit.name, KyoPageEdit);\n// import KyoPageBuilder from \"./builder/KyoPageBuilder\";\n// App.component(KyoPageBuilder.name, KyoPageBuilder);\n// import KyoWidgetHeadline from \"./widgets/KyoWidgetHeadline\";\n// App.component(KyoWidgetHeadline.name, KyoWidgetHeadline);\n// require('./configs/page');\n\nKyoto.ready(function (App) {\n  App.component(_pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__[\"default\"].name, _pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n});\n\nif (console && console.log) {\n  console.log('kyoto/page ready.');\n}\n\n//# sourceURL=webpack:///./resources/js/bootstrap.js?");
 
 /***/ }),
 
-/***/ "./resources/js/builder/KyoPageBuilder.jsx":
-/*!*************************************************!*\
-  !*** ./resources/js/builder/KyoPageBuilder.jsx ***!
-  \*************************************************/
+/***/ "./resources/js/pages/KyoPages2.vue":
+/*!******************************************!*\
+  !*** ./resources/js/pages/KyoPages2.vue ***!
+  \******************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _vue_babel_helper_vue_jsx_merge_props__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @vue/babel-helper-vue-jsx-merge-props */ \"./node_modules/@vue/babel-helper-vue-jsx-merge-props/dist/helper.js\");\n/* harmony import */ var _vue_babel_helper_vue_jsx_merge_props__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_vue_babel_helper_vue_jsx_merge_props__WEBPACK_IMPORTED_MODULE_0__);\n\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageWidgets',\n  props: {\n    value: {\n      \"default\": function _default() {\n        return [];\n      }\n    }\n  },\n  data: function data() {\n    var widgets = [{\n      id: Nano.UUID(),\n      value: {\n        title: ''\n      },\n      type: 'headline',\n      name: 'test1'\n    }, {\n      id: Nano.UUID(),\n      value: {\n        title: ''\n      },\n      type: 'headline',\n      name: 'test2'\n    }];\n    return {\n      widgets: widgets\n    };\n  },\n  methods: {\n    transformDrop: function transformDrop(source) {\n      return Nano.Obj.assign(source, {\n        id: Nano.UUID()\n      });\n    },\n    allowDrop: function allowDrop(source, target, strategy) {\n      return !this.value.length || strategy !== 'root';\n    }\n  },\n  renderWidgetNode: function renderWidgetNode(props) {\n    var h = this.$createElement;\n    return h(\"div\", {\n      \"class\": \"kyo-widgets-widget-item\"\n    }, [props.value.type]);\n  },\n  renderBuilderNode: function renderBuilderNode(props) {\n    var h = this.$createElement;\n    return h(\"NForm\", {\n      \"class\": \"kyo-widgets-builder-item\"\n    }, [h(\"div\", {\n      \"class\": \"kyo-widgets-builder-item__head\",\n      \"attrs\": {\n        \"draggable\": \"true\"\n      }\n    }, [h(\"div\", {\n      \"class\": \"grid grid--row grid--middle grid--10\"\n    }, [h(\"div\", {\n      \"class\": \"col--auto col--left\"\n    }, [props.value.type]), h(\"div\", {\n      \"class\": \"col--auto\"\n    }, [h(\"NButton\", {\n      \"attrs\": {\n        \"size\": \"mini\",\n        \"type\": \"info\",\n        \"square\": true,\n        \"icon\": this.icons.create\n      },\n      \"on\": {\n        \"click\": props.copy\n      }\n    })]), h(\"div\", {\n      \"class\": \"col--auto\"\n    }, [h(\"NButton\", {\n      \"attrs\": {\n        \"size\": \"mini\",\n        \"type\": \"danger\",\n        \"square\": true,\n        \"icon\": this.icons.times\n      },\n      \"on\": {\n        \"click\": props.remove\n      }\n    })])])]), h(\"div\", {\n      \"class\": \"kyo-widgets-builder-item__body\"\n    }, [this.$render(\"KyoWidget\".concat(Nano.Str.ucfirst(props.value.type)))])]);\n  },\n  render: function render($render) {\n    var _this = this;\n\n    var h = arguments[0];\n    this.$render = $render;\n    var builderProps = {\n      items: this.value,\n      group: ['page-widget'],\n      handle: true,\n      itemHeight: 0,\n      viewportHeight: true,\n      removeNode: false,\n      insertNode: true,\n      useRenderCache: false,\n      keyEvents: false,\n      threshold: 0,\n      allowCurrent: false,\n      transformDrop: this.transformDrop,\n      allowDrop: this.allowDrop,\n      renderNode: this.ctor('renderBuilderNode')\n    };\n    var widgetProps = {\n      items: this.widgets,\n      group: ['page-widget'],\n      itemHeight: 0,\n      viewportHeight: true,\n      removeNode: false,\n      insertNode: false,\n      useRenderCache: false,\n      keyEvents: false,\n      threshold: 0,\n      allowCurrent: false,\n      allowDrop: false,\n      renderNode: this.ctor('renderWidgetNode')\n    };\n    var builderEvents = {\n      'input': function input(value) {\n        return _this.$emit('input', value);\n      }\n    };\n    return h(\"div\", {\n      \"class\": \"kyo-widgets\"\n    }, [h(\"div\", {\n      \"class\": \"kyo-widgets__inner\"\n    }, [h(\"NDraglist\", _vue_babel_helper_vue_jsx_merge_props__WEBPACK_IMPORTED_MODULE_0___default()([{\n      \"class\": \"kyo-widgets__widget\",\n      \"ref\": \"widget\"\n    }, {\n      \"props\": widgetProps\n    }])), h(\"NDraglist\", _vue_babel_helper_vue_jsx_merge_props__WEBPACK_IMPORTED_MODULE_0___default()([{\n      \"class\": \"kyo-widgets__builder\",\n      \"ref\": \"builder\"\n    }, {\n      \"props\": builderProps\n    }, {}, {\n      \"on\": builderEvents\n    }])), h(\"div\", {\n      \"class\": \"kyo-widgets__extra\"\n    }, [this.$slots[\"default\"]])])]);\n  }\n});\n\n//# sourceURL=webpack:///./resources/js/builder/KyoPageBuilder.jsx?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPages2.vue?vue&type=template&id=7c630b54 */ \"./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54\");\n/* harmony import */ var _KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPages2.vue?vue&type=script&lang=js */ \"./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/pages/KyoPages2.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
 
 /***/ }),
 
-/***/ "./resources/js/configs/page.js":
-/*!**************************************!*\
-  !*** ./resources/js/configs/page.js ***!
-  \**************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-eval("Nano.Data.set('kyoto/page::page', {\n  'NFormGroup:00': {\n    props: {\n      legend: Nano.Locale.trans('Page'),\n      icon: 'fa fa-file'\n    },\n    content: {\n      'div:00': {\n        \"class\": 'grid grid--row grid--wrap grid--30',\n        content: {\n          'div:00': {\n            \"class\": 'col--1-1 col--1-2@sm',\n            content: {\n              'NFormItem:00': {\n                props: {\n                  label: Nano.Locale.trans('Page'),\n                  prop: 'option.page_id'\n                },\n                content: {\n                  'NInput:00': {\n                    model: {\n                      path: 'option.page_id'\n                    }\n                  }\n                }\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n});\n\n//# sourceURL=webpack:///./resources/js/configs/page.js?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageCreate.vue":
-/*!**********************************************!*\
-  !*** ./resources/js/pages/KyoPageCreate.vue ***!
-  \**********************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPageCreate.vue?vue&type=template&id=3f209878& */ \"./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878&\");\n/* harmony import */ var _KyoPageCreate_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPageCreate.vue?vue&type=script&lang=js& */ \"./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _KyoPageCreate_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js&":
-/*!***********************************************************************!*\
-  !*** ./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js& ***!
-  \***********************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageCreate_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPageCreate.vue?vue&type=script&lang=js& */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageCreate_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878&":
-/*!*****************************************************************************!*\
-  !*** ./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878& ***!
-  \*****************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPageCreate.vue?vue&type=template&id=3f209878& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=3f209878&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageCreate_vue_vue_type_template_id_3f209878___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageEdit.vue":
-/*!********************************************!*\
-  !*** ./resources/js/pages/KyoPageEdit.vue ***!
-  \********************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPageEdit.vue?vue&type=template&id=3a712f8a& */ \"./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a&\");\n/* harmony import */ var _KyoPageEdit_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPageEdit.vue?vue&type=script&lang=js& */ \"./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _KyoPageEdit_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js&":
-/*!*********************************************************************!*\
-  !*** ./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js& ***!
-  \*********************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageEdit_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPageEdit.vue?vue&type=script&lang=js& */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageEdit_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a&":
-/*!***************************************************************************!*\
-  !*** ./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a& ***!
-  \***************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPageEdit.vue?vue&type=template&id=3a712f8a& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=3a712f8a&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPageEdit_vue_vue_type_template_id_3a712f8a___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPages.vue":
-/*!*****************************************!*\
-  !*** ./resources/js/pages/KyoPages.vue ***!
-  \*****************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPages.vue?vue&type=template&id=6c1de27c& */ \"./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c&\");\n/* harmony import */ var _KyoPages_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPages.vue?vue&type=script&lang=js& */ \"./resources/js/pages/KyoPages.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _KyoPages_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/pages/KyoPages.vue?vue&type=script&lang=js&":
+/***/ "./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js":
 /*!******************************************************************!*\
-  !*** ./resources/js/pages/KyoPages.vue?vue&type=script&lang=js& ***!
+  !*** ./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js ***!
   \******************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPages_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPages.vue?vue&type=script&lang=js& */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPages_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages2.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
 
 /***/ }),
 
-/***/ "./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c&":
+/***/ "./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54":
 /*!************************************************************************!*\
-  !*** ./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c& ***!
+  !*** ./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54 ***!
   \************************************************************************/
-/*! exports provided: render, staticRenderFns */
+/*! exports provided: render */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoPages.vue?vue&type=template&id=6c1de27c& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/pages/KyoPages.vue?vue&type=template&id=6c1de27c&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoPages_vue_vue_type_template_id_6c1de27c___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/widgets/KyoWidgetHeadline.vue":
-/*!****************************************************!*\
-  !*** ./resources/js/widgets/KyoWidgetHeadline.vue ***!
-  \****************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0& */ \"./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0&\");\n/* harmony import */ var _KyoWidgetHeadline_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoWidgetHeadline.vue?vue&type=script&lang=js& */ \"./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _KyoWidgetHeadline_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js&":
-/*!*****************************************************************************!*\
-  !*** ./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js& ***!
-  \*****************************************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoWidgetHeadline_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoWidgetHeadline.vue?vue&type=script&lang=js& */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoWidgetHeadline_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
-
-/***/ }),
-
-/***/ "./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0&":
-/*!***********************************************************************************!*\
-  !*** ./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0& ***!
-  \***********************************************************************************/
-/*! exports provided: render, staticRenderFns */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=1c510cd0&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_KyoWidgetHeadline_vue_vue_type_template_id_1c510cd0___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages2.vue?vue&type=template&id=7c630b54 */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
 
 /***/ }),
 
